[MOD] Competency 타입 공용 파일로 분리 및 참조 통일
